# Small optimization for FromLabelAdaptersToLabelsWithCopy